+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 - Add single source of truth for notifications.
+- Update homepage to use single route for tabs.
 
 ### Fixed
 - Added model toast persisted in route state.